Extract Primary Details Component
